feat: [ft-wave1-cli-json-result-004] - Consume three migration-ledger rows and close cell verification gates
