+++ docs/temp/functional-tests-expansion/migration-ledger-inventory.json
@@ -631,16 +631,6 @@
       "deletion_only_batch": "n/a"
     },
     {
-      "source_path": "tests/functional/cli/factory_run/output/failure_test.go",
-      "package": "you-agent-factory/tests/functional/cli/factory_run/output",
-      "scenario": "TestOutputSelectionFailsBeforeProductActivation",
-      "lane": "short",
-      "destination": "tests/functional/transport/cli/output/json_result_test.go",
-      "catch_all": "none",
-      "specialty_targets": "none",
-      "deletion_only_batch": "n/a"
-    },
-    {
       "source_path": "tests/functional/cli/factory_run/output/output_test.go",
       "package": "you-agent-factory/tests/functional/cli/factory_run/output",
       "scenario": "TestSuccessfulInvocationOutputModes",
@@ -4101,6 +4091,36 @@
       "deletion_only_batch": "n/a"
     },
     {
+      "source_path": "tests/functional/orchestration/javascript/workers/overrides_test.go",
+      "package": "you-agent-factory/tests/functional/orchestration/javascript/workers",
+      "scenario": "TestJavaScriptChildrenSelectDifferentProvidersAndModels",
+      "lane": "short",
+      "destination": "tests/functional/orchestration/javascript/workers/overrides_test.go",
+      "catch_all": "none",
+      "specialty_targets": "none",
+      "deletion_only_batch": "n/a"
+    },
+    {
+      "source_path": "tests/functional/orchestration/javascript/workers/overrides_test.go",
+      "package": "you-agent-factory/tests/functional/orchestration/javascript/workers",
+      "scenario": "TestJavaScriptMockWorkersReplaceOnlyNamedChildren",
+      "lane": "short",
+      "destination": "tests/functional/orchestration/javascript/workers/overrides_test.go",
+      "catch_all": "none",
+      "specialty_targets": "none",
+      "deletion_only_batch": "n/a"
+    },
+    {
+      "source_path": "tests/functional/orchestration/javascript/workers/overrides_test.go",
+      "package": "you-agent-factory/tests/functional/orchestration/javascript/workers",
+      "scenario": "TestJavaScriptUnknownWorkerOverrideFailsActionably",
+      "lane": "short",
+      "destination": "tests/functional/orchestration/javascript/workers/overrides_test.go",
+      "catch_all": "none",
+      "specialty_targets": "none",
+      "deletion_only_batch": "n/a"
+    },
+    {
       "source_path": "tests/functional/smoke/cli_named_deep_research_smoke_test.go",
       "package": "you-agent-factory/tests/functional/smoke",
       "scenario": "TestNamedDeepResearchCLI_DefaultInvocationReturnsLeadSynthesis",
@@ -4836,6 +4856,46 @@
       "scenario": "TestCLISubmitUnavailableServer",
       "lane": "short",
       "destination": "tests/functional/transport/cli/commands/submit_wiring_test.go",
+      "catch_all": "none",
+      "specialty_targets": "none",
+      "deletion_only_batch": "n/a"
+    },
+    {
+      "source_path": "tests/functional/transport/cli/output/json_result_test.go",
+      "package": "you-agent-factory/tests/functional/transport/cli/output",
+      "scenario": "TestCLIJSONContainsNoPrivateRuntimeFields",
+      "lane": "short",
+      "destination": "tests/functional/transport/cli/output/json_result_test.go",
+      "catch_all": "none",
+      "specialty_targets": "none",
+      "deletion_only_batch": "n/a"
+    },
+    {
+      "source_path": "tests/functional/transport/cli/output/json_result_test.go",
+      "package": "you-agent-factory/tests/functional/transport/cli/output",
+      "scenario": "TestCLIJSONFailureRemainsValidJSON",
+      "lane": "short",
+      "destination": "tests/functional/transport/cli/output/json_result_test.go",
+      "catch_all": "none",
+      "specialty_targets": "none",
+      "deletion_only_batch": "n/a"
+    },
+    {
+      "source_path": "tests/functional/transport/cli/output/json_result_test.go",
+      "package": "you-agent-factory/tests/functional/transport/cli/output",
+      "scenario": "TestCLIJSONOutputSelectionFailsBeforeProductActivation",
+      "lane": "short",
+      "destination": "tests/functional/transport/cli/output/json_result_test.go",
+      "catch_all": "none",
+      "specialty_targets": "none",
+      "deletion_only_batch": "n/a"
+    },
+    {
+      "source_path": "tests/functional/transport/cli/output/json_result_test.go",
+      "package": "you-agent-factory/tests/functional/transport/cli/output",
+      "scenario": "TestCLIJSONSuccessDecodesToPublicInvocationResult",
+      "lane": "short",
+      "destination": "tests/functional/transport/cli/output/json_result_test.go",
       "catch_all": "none",
       "specialty_targets": "none",
       "deletion_only_batch": "n/a"
@@ -6359,26 +6419,26 @@
       "sessionparity": 13,
       "sessions": 2,
       "smoke": 84,
-      "transport": 34,
+      "transport": 37,
       "work": 2,
       "workers": 40,
       "workflow": 64,
       "workstations": 7
     },
-    "customer_top_level_Test_scenarios": 614,
+    "customer_top_level_Test_scenarios": 620,
     "files_with_customer_Test": 200,
     "functional_test_go_files": 251,
     "helper_only_non_customer_files": 51,
     "internal_harness_Test_functions": 15,
     "internal_harness_test_files": 5,
     "lane_functionallong": 118,
-    "lane_short": 496,
+    "lane_short": 502,
     "by_full_package": {
       "you-agent-factory/tests/functional/acceptance": 14,
       "you-agent-factory/tests/functional/bootstrap_portability": 25,
       "you-agent-factory/tests/functional/bootstrap_portability/metadata_contract": 2,
       "you-agent-factory/tests/functional/cli/dynamic_workflows": 2,
-      "you-agent-factory/tests/functional/cli/factory_run/output": 3,
+      "you-agent-factory/tests/functional/cli/factory_run/output": 2,
       "you-agent-factory/tests/functional/cli/factory_run/yaml_io_parity": 6,
       "you-agent-factory/tests/functional/cli/input_contract": 25,
       "you-agent-factory/tests/functional/cli/mcp_resume": 8,
@@ -6397,6 +6457,7 @@
       "you-agent-factory/tests/functional/orchestration/javascript/durability": 3,
       "you-agent-factory/tests/functional/orchestration/javascript/events": 2,
       "you-agent-factory/tests/functional/orchestration/javascript/loading": 2,
+      "you-agent-factory/tests/functional/orchestration/javascript/workers": 3,
       "you-agent-factory/tests/functional/orchestration/petri/dispatch": 3,
       "you-agent-factory/tests/functional/product/init_setup": 10,
       "you-agent-factory/tests/functional/providers": 39,
@@ -6410,7 +6471,7 @@
       "you-agent-factory/tests/functional/sessions/cli": 2,
       "you-agent-factory/tests/functional/smoke": 82,
       "you-agent-factory/tests/functional/transport/cli/commands": 4,
-      "you-agent-factory/tests/functional/transport/cli/output": 3,
+      "you-agent-factory/tests/functional/transport/cli/output": 7,
       "you-agent-factory/tests/functional/transport/cli/parameters": 3,
       "you-agent-factory/tests/functional/transport/docs": 1,
       "you-agent-factory/tests/functional/transport/http": 1,
@@ -6509,15 +6570,15 @@
     }
   },
   "completeness_audit": {
-    "audited_at_utc": "2026-07-27T15:40:00Z",
-    "story": "ft-wave1-js-policy-denied-mergeability",
-    "live_customer_scenarios": 614,
-    "ledger_rows": 614,
+    "audited_at_utc": "2026-07-27T16:05:00Z",
+    "story": "ft-wave1-cli-json-result-mergeability-6",
+    "live_customer_scenarios": 620,
+    "ledger_rows": 620,
     "unmapped_scenarios": 0,
     "checklist_destinations": 435,
     "wrong_layer_destinations": 69,
     "invalid_destinations": 0,
-    "lane_short": 496,
+    "lane_short": 502,
     "lane_functionallong": 118,
     "specialty_targets_documented": 11,
     "deletion_only_batches": 52,
