+++ tests/functional/cli/factory_run/output/failure_test.go
@@ -1,43 +1,19 @@
 package output
 
 import (
-	"context"
 	"encoding/json"
 	"io"
 	"os"
 	"strings"
-	"sync/atomic"
 	"testing"
 
-	platformhttpserver "github.com/portpowered/infinite-you/pkg/platform/httpserver"
 	serviceedges "github.com/portpowered/infinite-you/pkg/services/edges"
-	factorysessions "github.com/portpowered/infinite-you/pkg/services/factory_sessions"
 	"github.com/portpowered/infinite-you/pkg/services/workers"
-	runcli "github.com/portpowered/infinite-you/pkg/transports/cli/run"
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 	"github.com/portpowered/infinite-you/tests/functional/internal/support"
 )
 
 func TestInvocationFailureOutputContracts(t *testing.T) {
-	t.Run("failure before terminal response leaves stdout empty", func(t *testing.T) {
-		result := executeFailureInvocation(t, []string{
-			"you", "--json", "run", "--named", "@you/missing", "--no-record",
-			"deterministic pre-session failure",
-		}, "", nil)
-
-		if result.err == nil {
-			t.Fatal("Process.Execute error = nil, want pre-session invocation failure")
-		}
-		if result.stdout != "" {
-			t.Fatalf("stdout = %q, want no terminal record", result.stdout)
-		}
-		response := decodeSingleErrorResponse(t, result.stderr)
-		if response.Code != factoryapi.ErrorResponseCode(runcli.InvocationErrorCodeFailed) ||
-			response.Family != factoryapi.ErrorFamilyInternalServerError {
-			t.Fatalf("ErrorResponse = %#v", response)
-		}
-	})
-
 	t.Run("terminal failure emits failed result and standard error", func(t *testing.T) {
 		result := executeFailureInvocation(t, []string{
 			"you", "--json", "run", "--named", goalFactoryName, "--no-record",
@@ -112,71 +88,6 @@ func TestInvocationFailureOutputContracts(t *testing.T) {
 			t.Fatalf("canonical failed dispatch does not precede terminal invocation outcome\nstdout:\n%s", result.stdout)
 		}
 	})
-}
-
-// TestOutputSelectionFailsBeforeProductActivation proves invalid output choices remain side-effect free.
-func TestOutputSelectionFailsBeforeProductActivation(t *testing.T) {
-	tests := []struct {
-		name     string
-		args     []string
-		wantCode factoryapi.ErrorResponseCode
-	}{
-		{
-			name:     "quiet and global JSON",
-			args:     []string{"you", "--json", "run", "--quiet"},
-			wantCode: runcli.InvocationOutputConflictCode,
-		},
-		{
-			name:     "quiet and explicit output",
-			args:     []string{"you", "run", "--quiet", "--output", "primary"},
-			wantCode: runcli.InvocationOutputConflictCode,
-		},
-		{
-			name:     "unsupported explicit output",
-			args:     []string{"you", "run", "--output", "provider-chunks"},
-			wantCode: runcli.InvocationOutputUnsupportedCode,
-		},
-	}
-	for _, test := range tests {
-		test := test
-		t.Run(test.name, func(t *testing.T) {
-			var effects atomic.Int32
-			edges := serviceedges.Edges{
-				APIServerStarter: func(context.Context, platformhttpserver.StartRequest) error {
-					effects.Add(1)
-					return nil
-				},
-				BrowserOpener: func(context.Context, string) error {
-					effects.Add(1)
-					return nil
-				},
-				RuntimeHostObserver: func(factorysessions.RuntimeHostBinding) {
-					effects.Add(1)
-				},
-				FactorySessionIDGenerator: func() string {
-					effects.Add(1)
-					return "unexpected-session"
-				},
-			}
-			inputs := support.FakeInputs(t.Context(), test.args)
-			inputs.Input.WorkingDirectory = t.TempDir()
-
-			err := support.BuildProcess(t, edges).Execute(inputs.Input)
-			if err == nil {
-				t.Fatal("Process.Execute error = nil, want output-selection failure")
-			}
-			if inputs.Stdout() != "" {
-				t.Fatalf("stdout = %q, want empty", inputs.Stdout())
-			}
-			response := decodeSingleErrorResponse(t, inputs.Stderr())
-			if response.Code != test.wantCode || response.Family != factoryapi.ErrorFamilyBadRequest {
-				t.Fatalf("ErrorResponse = %#v, want code %s", response, test.wantCode)
-			}
-			if effects.Load() != 0 {
-				t.Fatalf("product effects = %d, want 0", effects.Load())
-			}
-		})
-	}
 }
 
 func rejectingGoalWorker() *workers.MockWorkersConfig {
